Use camera for aspect ratio correction instead of model matrix

+++ surface.h
@@ -16,7 +16,7 @@ struct Surface {
 	bool is_toplevel;
 
 	mat4 matrix;
-	int x, y, width, height;
+	double x, y, width, height;
 
 	// Set these and calc_matrices will do the rest. Rotations in radians, speeds in radians per frame.
 	double x_rot, y_rot, z_rot;
